refactor tests and package
