+++ cognite/neat/graph/extractors/__init__.py
@@ -1,4 +1,14 @@
-from . import graph, graph_capturing_sheet
-from .graph_store import NeatGraphStore, drop_graph_store
+from cognite.neat.graph.extractors import graph
+from cognite.neat.graph.extractors.graph_sheet_to_graph import (
+    extract_graph_from_sheet,
+    read_graph_excel_file_to_table_by_name,
+)
+from cognite.neat.graph.extractors.graph_store import NeatGraphStore, drop_graph_store
 
-__all__ = ["graph", "NeatGraphStore", "graph_capturing_sheet", "drop_graph_store"]
+__all__ = [
+    "graph",
+    "NeatGraphStore",
+    "extract_graph_from_sheet",
+    "read_graph_excel_file_to_table_by_name",
+    "drop_graph_store",
+]
