+++ cognite/neat/graph/loaders/__init__.py
@@ -1,7 +1,6 @@
-from cognite.neat.graph.loaders.cdfcore.labels import upload_labels
-from .cdfcore.rdf_to_assets import rdf2assets, upload_assets, categorize_assets
-from .cdfcore.rdf_to_relationships import rdf2relationships, categorize_relationships, upload_relationships
-from .graph_sheet_to_graph import sheet2triples
+from cognite.neat.graph.loaders.core.labels import upload_labels
+from .core.rdf_to_assets import rdf2assets, upload_assets, categorize_assets
+from .core.rdf_to_relationships import rdf2relationships, categorize_relationships, upload_relationships
 
 __all__ = [
     "rdf2relationships",
@@ -11,5 +10,4 @@ __all__ = [
     "categorize_relationships",
     "upload_relationships",
     "upload_labels",
-    "sheet2triples",
 ]
