+++ cognite/neat/graph/loaders/core/rdf_to_assets.py
@@ -16,7 +16,7 @@ from rdflib import Graph, Namespace
 from rdflib.term import URIRef
 
 from cognite.neat.app.api.configuration import EXCLUDE_PATHS
-from cognite.neat.graph.loaders.cdfcore.models import AssetTemplate
+from cognite.neat.graph.loaders.core.models import AssetTemplate
 from cognite.neat.graph.extractors.graph_store import NeatGraphStore
 from cognite.neat.rules.models import Property, TransformationRules
 from cognite.neat.utils.utils import chunker, datetime_utc_now, remove_namespace, retry_decorator
