+++ cognite/neat/rules/exporter/core/__init__.py
@@ -1,3 +1,5 @@
+"""Exporters dealing with CDF Core"""
+
 from .rules2labels import get_labels
 
 __all__ = ["get_labels"]
