+++ cognite/neat/workflows/base_workflows/graphs_and_rules.py
@@ -41,7 +41,7 @@ class GraphsAndRulesBaseWorkflow(BaseWorkflow):
         else:
             cdf_store.load_rules_file_from_cdf(self.cdf_client, version)
 
-        self.transformation_rules = rules.load_rules_from_excel_file(rules_file_path)
+        self.transformation_rules = rules.parse_rules_from_excel_file(rules_file_path)
         self.dataset_id = self.transformation_rules.metadata.data_set_id
         logging.info(f"Loaded prefixes {str(self.transformation_rules.prefixes)} rules from {rules_file_path.name!r}.")
         output_text = f"Loaded {len(self.transformation_rules.properties)} rules"
