+++ tests/graph/extractors/mocks/test_mock_graph.py
@@ -1,5 +1,5 @@
 from cognite.neat.graph import loaders, extractors
-from cognite.neat.graph.loaders.cdfcore import rdf_to_relationships
+from cognite.neat.graph.loaders.core import rdf_to_relationships
 from cognite.neat.graph.extractors.mocks import generate_triples
 from cognite.neat.rules.models import TransformationRules
 from cognite.neat.utils.utils import add_triples, remove_namespace
