+++ tests/graph/loaders/test_validator.py
@@ -1,4 +1,4 @@
-from cognite.neat.graph.loaders.cdfcore.rdf_to_assets import rdf2assets
+from cognite.neat.graph.loaders.core.rdf_to_assets import rdf2assets
 from cognite.neat.graph.extractors import NeatGraphStore
 from cognite.neat.graph.loaders.validator import validate_asset_hierarchy
