+++ tests/rules/exporter/test_rules2triples.py
