+++ tests/rules/importer/test_ontology2excel.py
