+++ tests/rules/test_analysis.py
